fix(81): address review feedback — shared auth, CORS fix, stricter grants
Review feedback addressed:

1. Extract duplicated authenticateRequest() to _shared/auth.ts with
   case-insensitive Bearer scheme parsing via regex (fixes fragile
   header parsing flagged by Qodo).

2. Fix CORS race condition: replace module-level mutable corsHeaders
   with per-request const from getCorsHeaders() passed through all
   handlers (flagged by Claude review as medium priority).

3. Strengthen get_indexed_recording_count grant: REVOKE FROM PUBLIC
   first, then re-GRANT only to authenticated — ensures no broader
   grants leak through (flagged by Qodo).
